separate file sys and environment functions
