+++ Top/csound.c
@@ -57,7 +57,6 @@
 #include "cs_par_base.h"
 #include "cs_par_orc_semantics.h"
 #include "namedins.h"
-// #include "cs_par_dispatch.h"
 #include "find_opcode.h"
 
 #if defined(linux) || defined(__HAIKU__) || defined(__EMSCRIPTEN__) ||         \
